Test 2

+++ my-app/pages/api/products/[id].js
@@ -7,7 +7,9 @@ export default async function handler(req, res) {
     const { 
         method, 
         query: { id }, 
+        cookies
     } = req;
+    const token = cookies.token
 
    await dbConnect();
 
@@ -35,10 +37,10 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
-    // if(!token || token !== process.env.token){
-    //   return res.status(401).json("Not authenticated!")
+    if(!token || token !== process.env.token){
+      return res.status(401).json("Not authenticated!")
    
-    // }
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted!");
